different flow for no url state
If you visit the site directly, it flips the order of the columns so
that when/where show up first with deets after. The idea is to hopefully
have the user realize that the inputs cause the deets table output.
Additionally, a link emoji along is shown next to "The Deets" in edit
mode so that it's more obvious that the URL is updating.